% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal_theme.R
\name{personal_theme}
\alias{personal_theme}
\title{Tema personalizado com definições para utilizar em gráficos no R}
\usage{
personal_theme(...)
}
\arguments{
\item{...}{Elementos do \code{ggplot2} que podem ser passados dentro da função
\code{ggplot2::theme()}}
}
\description{
Tema do \code{ggplot2} que particularmente gosto de utilizar quando estou
construindo gráficos.
}
