% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_custom.R
\name{theme_fabr}
\alias{theme_fabr}
\title{Tema personalizado com definições para utilizar em gráficos no R}
\usage{
theme_fabr(...)
}
\arguments{
\item{...}{Elementos do \code{ggplot2} que podem ser passados dentro da função
\code{ggplot2::theme()}}
}
\description{
Tema do \code{ggplot2} que particularmente gosto de utilizar quando estou
construindo gráficos.
}
