% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_data.R
\name{tidy_scopus}
\alias{tidy_scopus}
\title{Função para realizar limpeza dos dados do \code{Scopus}}
\usage{
tidy_scopus(path, pattern, ...)
}
\arguments{
\item{path}{Caminho da pasta na qual os arquivos oriundas do Scopus estão
localizados.}

\item{pattern}{Extensão do arquivo que será importado. Geralmente os arquivos
da \code{Scopus} vêm na extensão \code{.csv} (comma separated value).}

\item{...}{Relação de colunas que do arquivo do Scopus que o usuário deseja
selecionar.}
}
\description{
A função que realiza a importação e limpeza dos dados do \code{Scopus}
para realização de análises bibliométricas.
}
