% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wb_data.R
\name{tidy_wb}
\alias{tidy_wb}
\title{Função para realizar limpeza dos dados da \verb{Web of Science}}
\usage{
tidy_wb(path, pattern, ...)
}
\arguments{
\item{path}{Caminho da pasta na qual os arquivos oriundas da Web of Science
estão localizados.}

\item{pattern}{Extensão do arquivo que será importado. Geralmente os arquivos
da \verb{Web of Science} vêm na extensão \code{.xls} ou \code{.xlsx}, planilhas do Excel.}

\item{...}{Relação de colunas que do arquivo da Web of Science que o usuário
deseja selecionar.}
}
\description{
A função que realiza a importação e limpeza dos dados da \verb{Web of Science}
para realização de análises bibliométricas.
}
